+++ AsBuiltReport.Veeam.VBR/Src/Private/Diagram/New-AbrVeeamDiagram.ps1
@@ -374,11 +374,6 @@ function New-AbrVeeamDiagram {
 
     begin {
 
-        if ($psISE) {
-            Write-Error -Message 'You cannot run this script inside the PowerShell ISE. Please execute it from the PowerShell Command Window.'
-            break
-        }
-
         if ($DiagramType -eq 'Backup-to-CloudConnect-Tenant' -and ([string]::IsNullOrEmpty($TenantName) -eq $true)) {
             throw 'TenantName must be a used with the Backup-to-CloudConnect-Tenant diagram type.'
         }
@@ -387,27 +382,6 @@ function New-AbrVeeamDiagram {
             $PSCmdlet.MyInvocation.BoundParameters['Verbose'].IsPresent
         } else {
             $false
-        }
-
-        if ($EnableErrorDebug) {
-            $global:VerbosePreference = 'Continue'
-            $global:DebugPreference = 'Continue'
-        } else {
-            $global:VerbosePreference = 'SilentlyContinue'
-            $global:DebugPreference = 'SilentlyContinue'
-        }
-
-        #@tpcarman
-        # If Username and Password parameters used, convert specified Password to secure string and store in $Credential
-        if ($Username) {
-            if (-not $Password) {
-                # If the Password parameter is not provided, prompt for it securely
-                $SecurePassword = Read-Host "Password for user '$Username'" -AsSecureString
-            } else {
-                # If the Password parameter is provided, convert it to secure string
-                $SecurePassword = ConvertTo-SecureString $Password -AsPlainText -Force
-            }
-            $Credential = New-Object System.Management.Automation.PSCredential ($Username, $SecurePassword)
         }
 
         if (($Format -ne 'base64') -and !(Test-Path $OutputFolderPath)) {
@@ -433,37 +407,6 @@ function New-AbrVeeamDiagram {
             'Backup-to-CloudConnect-Tenant' { "Cloud Connect $TenantName Resources Diagram" }
             'Backup-to-HACluster' { 'High Availability Diagram' }
         }
-        if ($Format -ne 'Base64') {
-            Write-AbrColorOutput -Color 'Green' -String ("Please wait while the '{0}' is being generated." -f $MainGraphLabel)
-            Write-AbrColorOutput -Color 'White' -String ' - Please refer to the Veeam.Diagrammer github website for more detailed information about this project.'
-            Write-AbrColorOutput -Color 'White' -String ' - Documentation: https://github.com/rebelinux/Veeam.Diagrammer'
-            Write-AbrColorOutput -Color 'White' -String ' - Issues or bug reporting: https://github.com/rebelinux/Veeam.Diagrammer/issues'
-            Write-AbrColorOutput -Color 'White' -String ' - This project is community maintained and has no sponsorship from Veeam, its employees or any of its affiliates.'
-
-
-            # Check the version of the dependency modules
-            if ($UpdateCheck) {
-                $ModuleArray = @('Veeam.Diagrammer', 'Diagrammer.Core')
-
-                foreach ($Module in $ModuleArray) {
-                    try {
-                        $InstalledVersion = Get-Module -ListAvailable -Name $Module -ErrorAction SilentlyContinue | Sort-Object -Property Version -Descending | Select-Object -First 1 -ExpandProperty Version
-
-                        if ($InstalledVersion) {
-                            Write-AbrColorOutput -Color 'White' -String " - $Module module v$($InstalledVersion.ToString()) is currently installed."
-                            $LatestVersion = Find-Module -Name $Module -Repository PSGallery -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Version
-                            if ($InstalledVersion -lt $LatestVersion) {
-                                Write-AbrColorOutput -Color 'Red' -String "  - $Module module v$($LatestVersion.ToString()) is available."
-                                Write-AbrColorOutput -Color 'Red' -String "  - Run 'Update-Module -Name $Module -Force' to install the latest version."
-                            }
-                        }
-                    } catch {
-                        Write-Error $_.Exception.Message
-                    }
-                }
-            }
-        }
-
         $IconDebug = $false
 
         if ($DraftMode) {
@@ -518,16 +461,12 @@ function New-AbrVeeamDiagram {
             }
         }
 
-        $script:NewIcons = $false
+        $script:NewIcons = $NewIcons.IsPresent
 
         $RootPath = Split-Path (Split-Path $PSScriptRoot -Parent) -Parent
         $IconPath = Join-Path "$RootPath\Tools" 'icons'
-        # $ImagePath = Join-Path $RootPath 'Private\Diagram\Images.ps1'
-
-        # . $ImagePath
 
         if ($DiagramType -eq 'Backup-Infrastructure') {
-
             $Dir = 'TB'
         } else {
             $Dir = switch ($Direction) {
@@ -730,15 +669,10 @@ function New-AbrVeeamDiagram {
             #Export Diagram
             foreach ($OutputFormat in $Format) {
 
-                $OutputDiagram = Export-AbrDiagram -GraphObj ($diGraph | Select-String -Pattern '"([A-Z])\w+"\s\[label="";style="invis";shape="point";]' -NotMatch) -ErrorDebug $EnableErrorDebug -Format $OutputFormat -Filename $Filename -OutputFolderPath $OutputFolderPath -WaterMarkText $WaterMarkText -WaterMarkColor $WaterMarkColor -IconPath $IconPath -Verbose:$Verbose -Rotate $Rotate
+                $OutputDiagram = Export-AbrDiagram -GraphObj ($diGraph | Select-String -Pattern '"([A-Z])\w+"\s\[label="";style="invis";shape="point";]' -NotMatch) -ErrorDebug $EnableErrorDebug -Format $OutputFormat -Filename $Filename -OutputFolderPath $OutputFolderPath -WaterMarkText $WaterMarkText -WaterMarkColor $WaterMarkColor -IconPath $IconPath -Verbose:$Verbose
 
                 if ($OutputDiagram) {
-                    if ($OutputFormat -ne 'Base64') {
-                        # If not Base64 format return image path
-                        Write-AbrColorOutput -Color 'White' -String ("Diagrammer diagram '{0}' has been saved to '{1}'" -f $OutputDiagram.Name, $OutputDiagram.Directory)
-                    } else {
-                        Write-PScriboMessage 'Displaying Base64 string'
-                        # Return Base64 string
+                    if ($OutputFormat -eq 'Base64') {
                         $OutputDiagram
                     }
                 }